Add metadata endpoint  to the fhir data save endpoint
